+++ dev/src/mergePCodeWords.h
@@ -34,14 +34,15 @@
          integer.  The first_bit argument defines the most significant
          bit of the merged word.  The bits are numbered from 0 - MAX_BIT-1.
 */
+#include <stdint.h>
 #ifndef MERGEPCODEWORDS_H
 #define MERGEPCODEWORDS_H
 
-   inline unsigned long merge( unsigned long w1,
-                        unsigned long w2,
+   inline uint32_t merge( uint32_t w1,
+                        uint32_t w2,
                         int first_bit )
    {
-      unsigned long outword = w1;
+      uint32_t outword = w1;
       if (first_bit==0) return(outword);
       
       outword <<= first_bit;
